rename: WritePage component로로 분리
